Finished team4 activity
