+++ src/public/partials/header.html
@@ -1,10 +1,9 @@
 <div class="logo">
   <img src="/images/noun_Tent_2517.svg" alt="tent image for logo">
-  <a href="index.html"> Sleep<span class="highlight">Outside</span></a>
+  <a href="/index.html"> Sleep<span class="highlight">Outside</span></a>
 </div>
 <div class="cart">
-  <a href="cart/index.html">
-
+  <a href="/cart/index.html">
     <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 100 100">
       <path
         d="M18.9 32.6c1.1 2.4 2.5 3.3 5.4 3.3 1.6 0 3.6-0.3 5.9-0.6 3.2-0.5 6.9-1 11.2-1 2.1 0 4.3 0.1 6.4 0.3 2.1 0.1 4.2 0.3 6.1 0.3 3.2 0 5.2-0.4 5.9-1.2 2.7-2.7 2.8-8.8 2.9-14.6 0.1-6.7 0.2-14.5 4.6-18.7 -0.5 0-1 0-1.6 0 -14.2 0-37.5 0-41.1 0C15.6 6.2 14.9 23.6 18.9 32.6z" />
